better colors / linetypes

+++ scripts/CNN/interactions_results.py
@@ -4,6 +4,7 @@ from imageio import imread
 import numpy as np
 from matplotlib.patches import Ellipse
 import os
+from os.path import join
 import tqdm
 import glob
 import train_interactions
@@ -17,7 +18,7 @@ def toarray(struct_array):
 
 def save_prediction_img(i, pred, out_filename, gt=None):
     fig = plt.figure()
-    im = imread(os.path.join(DATA_DIR, 'images_test', '%06d.jpg' % i))
+    im = imread(join(DATA_DIR, 'images_test', '%06d.jpg' % i))
     # im = X_test[i, :, :, :]
     # a = plt.subplot(111, aspect='equal')
     # a.imshow(im)
@@ -35,10 +36,10 @@ def save_prediction_img(i, pred, out_filename, gt=None):
 
     if gt is not None:
         ax.add_patch(Ellipse(gt[['ant1_x', 'ant1_y']][i], gt[i]['ant1_major'], gt[i]['ant1_minor'],
-                             angle=-gt[i]['ant1_angle'], edgecolor='magenta', facecolor='none', label='ant1 gt'))
+                             angle=-gt[i]['ant1_angle'], edgecolor='red', linestyle='dotted', facecolor='none', label='ant1 gt'))
         # ax.add_patch(Ellipse((pred[i, 5], pred[i, 6]), pred[i, 7], pred[i, 8], angle=pred[i, 9], edgecolor='blue', facecolor='none'))
         ax.add_patch(Ellipse(gt[['ant2_x', 'ant2_y']][i], gt[i]['ant2_major'], gt[i]['ant2_minor'],
-                             angle=-gt[i]['ant2_angle'], edgecolor='cyan', facecolor='none', label='ant2 gt'))
+                             angle=-gt[i]['ant2_angle'], edgecolor='blue', linestyle='dotted', facecolor='none', label='ant2 gt'))
 
     # ells[i].set_clip_box(a.bbox)
     # ells[i].set_alpha(0.5)
@@ -54,8 +55,8 @@ def save_prediction_img(i, pred, out_filename, gt=None):
 if __name__ == '__main__':
     # ROOT_DIR = '/Users/flipajs/Downloads/double_regions'
     # ROOT_DIR = '/Users/flipajs/Documents/wd/FERDA/cnn_exp'
-    ROOT_DIR = '/datagrid/personal/smidm1/ferda/iteractions/'
-    DATA_DIR = ROOT_DIR
+    EXPERIMENT_DIR = '/datagrid/personal/smidm1/ferda/iteractions/experiments/171129_2121'
+    DATA_DIR = '/datagrid/personal/smidm1/ferda/iteractions/'
 
     NAMES = 'ant1_x, ant1_y, ant1_major, ant1_minor, ant1_angle, ' \
             'ant2_x, ant2_y, ant2_major, ant2_minor, ant2_angle'
@@ -64,14 +65,14 @@ if __name__ == '__main__':
     # with h5py.File(DATA_DIR + '/imgs_inter_test.h5', 'r') as hf:
     #     X_test = hf['data'][:]
     #
-    with h5py.File(DATA_DIR + '/results_inter_test.h5', 'r') as hf:
+    with h5py.File(join(DATA_DIR, 'results_inter_test.h5'), 'r') as hf:
         y_test = np.core.records.fromarrays(hf['data'][:].transpose(), names=NAMES, formats=FORMATS)
 
-    with h5py.File(DATA_DIR + '/predictions.h5', 'r') as hf:
+    with h5py.File(join(EXPERIMENT_DIR, 'predictions.h5'), 'r') as hf:
         pred = np.core.records.fromarrays(hf['data'][:].transpose(), names=NAMES, formats=FORMATS)
         import pandas as pd
         df = pd.DataFrame(pred)
-        # df.to_csv(os.path.join(DATA_DIR, 'pred.csv'))
+        # df.to_csv(join(DATA_DIR, 'pred.csv'))
         df.describe()
 
     # delta = 45.0  # degrees
@@ -86,22 +87,26 @@ if __name__ == '__main__':
     xy2_error = np.linalg.norm(toarray(y_test[['ant2_x', 'ant2_y']]) - toarray(pred[['ant2_x', 'ant2_y']]), 2, axis=1)
     angle1_error, angle2_error = train_interactions.angle_absolute_error(toarray(y_test), toarray(pred), np)
 
-
-    for fn in glob.glob(os.path.join(ROOT_DIR, 'test_predictions', '*.png')):
-        os.remove(fn)
+    if not os.path.exists(EXPERIMENT_DIR):
+        os.mkdir(EXPERIMENT_DIR)
+    if not os.path.exists(join(EXPERIMENT_DIR, 'test_predictions')):
+        os.mkdir(join(EXPERIMENT_DIR, 'test_predictions'))
+    else:
+        for fn in glob.glob(join(EXPERIMENT_DIR, 'test_predictions', '*.png')):
+            os.remove(fn)
 
     for i in tqdm.tqdm(angle1_error.flatten().argsort()[::-1][:20]):
-        save_prediction_img(i, pred, os.path.join(ROOT_DIR, 'test_predictions', 'bad_angle1_%03d.png' % i), y_test)
+        save_prediction_img(i, pred, join(EXPERIMENT_DIR, 'test_predictions', 'bad_angle1_%03d.png' % i), y_test)
     for i in tqdm.tqdm(angle2_error.flatten().argsort()[::-1][:20]):
-        save_prediction_img(i, pred, os.path.join(ROOT_DIR, 'test_predictions', 'bad_angle2_%03d.png' % i), y_test)
+        save_prediction_img(i, pred, join(EXPERIMENT_DIR, 'test_predictions', 'bad_angle2_%03d.png' % i), y_test)
     for i in tqdm.tqdm(xy1_error.flatten().argsort()[::-1][:20]):
-        save_prediction_img(i, pred, os.path.join(ROOT_DIR, 'test_predictions', 'bad_xy1_%03d.png' % i), y_test)
+        save_prediction_img(i, pred, join(EXPERIMENT_DIR, 'test_predictions', 'bad_xy1_%03d.png' % i), y_test)
     for i in tqdm.tqdm(xy2_error.flatten().argsort()[::-1][:20]):
-        save_prediction_img(i, pred, os.path.join(ROOT_DIR, 'test_predictions', 'bad_xy2_%03d.png' % i), y_test)
+        save_prediction_img(i, pred, join(EXPERIMENT_DIR, 'test_predictions', 'bad_xy2_%03d.png' % i), y_test)
 
 
     for i in tqdm.tqdm(np.random.randint(0, len(pred), 50)):
-        save_prediction_img(i, pred, os.path.join(ROOT_DIR, 'test_predictions', 'random_%04d.png' % i), y_test)
+        save_prediction_img(i, pred, join(EXPERIMENT_DIR, 'test_predictions', 'random_%04d.png' % i), y_test)
         # plt.show()
 
     # create montages in fish shell:
